backend/pyproject.toml
[project]
name = "wm-predictor-backend"
version = "0.1.0"
description = "WM 2026 Predictor backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.4",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "numpy>=2.0",
    "scipy>=1.13",
    "pandas>=2.2",
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6"]

[tool.ruff]
line-length = 100
target-version = "py312"
